requirements.txt
django==1.11
django-registration-redux == 1.7
djangorestframework==3.6.3
wagtail == 1.12.*
